Move all utilities into prometheus.Prometheus
These were all prometheus-specific anyway, so this feels better from an
encapsulation perspective.
